// UI/CodePulse/src/app/Features/blog-post/models/blog-post-model.ts
import { Category } from "../../Category/models/categorymodel";

export interface BlogPost {
    id: string;
    title:string;
    shortDescription: string;
    content: string;
    featuredImageUrl: string;
    urlHandle: string;
    author: string;
    publishedDate: Date;
    isVisible: boolean;
    categories: Category[];
}